Data.Time.Millis

Description

Avro defines a specific logical type for time differences expessed in milliseconds. This module provides a type which wraps the DiffTime from the time library (which uses nanoseconds), offering a millisecond-based interface.

Documentation

newtype DiffTimeMs

Wrapper for time difference expressed in milliseconds

Constructors

DiffTimeMs

Fields

unDiffTimeMs :: DiffTime

diffTimeToMillis :: DiffTimeMs -> Integer

Obtain the underlying time in milliseconds from a DiffTimeMs.

millisToDiffTime :: Integer -> DiffTimeMs

Build a DiffTimeMs from an amount expressed in milliseconds.