Mu.Quasi.Avro

Description

This module turns schema definitions written in Avro IDL into Mu Schemas. We provide versions for writing the IDL inline (avro) and import it from a file (avroFile).

Note: as of now, only the JSON-based IDL format is supported, not the Java-like one.

Quasi-quoters for .avsc files

avro :: QuasiQuoter

Imports an avro definition written in-line as a Schema.

avroFile :: QuasiQuoter

Imports an avro definition from a file as a Schema.

Only for internal use

schemaFromAvroType :: Type -> Q Type

Turns a schema from Avro into a Template Haskell Type.